requirements.txt
fastapi
uvicorn[standard]
websockets
websocket-client
python-dotenv
pyotp
requests
pandas
NorenRestApiPy
